refactor: use simple ts
