+++ src/import-service/package.json
@@ -4,16 +4,22 @@
   "description": "aws import with S3",
   "main": "index.js",
   "scripts": {
-    "deploy": "serverless deploy",
-    "build:webpack": "serverless webpack"
+    "deploy": "tsc && serverless deploy",
+    "build:webpack": "serverless webpack",
+    "build": "tsc"
   },
   "author": "AAA",
   "license": "ISC",
   "dependencies": {
     "aws-sdk": "^2.899.0",
+    "csv-parser": "^3.0.0"
+  },
+  "devDependencies": {
     "copy-webpack-plugin": "^8.1.1",
-    "csv-parser": "^3.0.0",
     "serverless-webpack": "^5.4.2",
+    "ts-loader": "^9.1.1",
+    "tsc": "^1.20150623.0",
+    "typescript": "^4.2.4",
     "webpack": "^5.36.2",
     "webpack-cli": "^4.6.0",
     "webpack-node-externals": "^3.0.0"
