+++ src/import-service/serverless.yml
@@ -66,7 +66,7 @@ package:
   patterns:
     - "!*"
     - "!*/**"
-    - handler.js
+    - dist/**
   # include:
   #   - dist/**
   # exclude:
